// src/dorc-web/src/apis/dorc-api/models/ContainerApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface ContainerApiModel
 */
export interface ContainerApiModel {
    /**
     * @type {Array<string>}
     * @memberof ContainerApiModel
     */
    EnvironmentNames?: Array<string> | null;
    /**
     * @type {boolean}
     * @memberof ContainerApiModel
     */
    UserEditable?: boolean;
    /**
     * @type {number}
     * @memberof ContainerApiModel
     */
    Id?: number;
    /**
     * @type {string}
     * @memberof ContainerApiModel
     */
    Name?: string | null;
    /**
     * @type {string}
     * @memberof ContainerApiModel
     */
    Image?: string | null;
    /**
     * @type {string}
     * @memberof ContainerApiModel
     */
    Registry?: string | null;
    /**
     * @type {string}
     * @memberof ContainerApiModel
     */
    HostServerName?: string | null;
    /**
     * @type {string}
     * @memberof ContainerApiModel
     */
    Tags?: string | null;
}